fix: move expression simplification to its own file
